added user profile and refactored some code

+++ src/main/java/com/project/taskipro/service/UserServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -22,6 +23,8 @@ public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
 
     private final UserMapper userMapper;
+
+    private final PasswordEncoder passwordEncoder;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws ResponseStatusException {
@@ -55,29 +58,51 @@ public class UserServiceImpl implements UserService {
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Пользователь не найден!"));
     }
 
-    public UserResponseDto getUserDtoById(Long id) throws ResponseStatusException {
-        return userMapper.toUserResponseDto(userRepository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,  String.format("Пользователь с id %d не найден", id))));
+    public UserResponseDto getUserDtoById() throws ResponseStatusException {
+        return userMapper.toUserResponseDto(getCurrentUser());
     }
 
-    public void setUserFields(Long userId, UserFieldsDto request){
-        User user = getUserById(userId);
-        user.setFirstname(request.firstname());
-        user.setLastname(request.lastname());
-        user.setPassword(request.password());
+    public void setUserFields(UserFieldsDto request) {
+        User user = getCurrentUser();
+
+        request.username().ifPresent(user::setUsername);
+        request.firstname().ifPresent(user::setFirstname);
+        request.lastname().ifPresent(user::setLastname);
+
+        if (request.oldPassword().isPresent() || request.newPassword().isPresent()) {
+            if (request.oldPassword().isEmpty() || request.newPassword().isEmpty()) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Для смены пароля нужно указать и старый, и новый пароль");
+            }
+
+            String oldPassword = request.oldPassword().get();
+            if (!passwordEncoder.matches(oldPassword, user.getPassword())) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Неверный текущий пароль");
+            }
+
+            String newPassword = request.newPassword().get();
+            user.setPassword(passwordEncoder.encode(newPassword));
+        }
 
         userRepository.save(user);
     }
 
-    public void deleteUserById(Long id){
-        User user = getUserById(id);
+    public void deleteUserById(UserFieldsDto request){
+        User user = getCurrentUser();
 
+        String oldPassword = request.oldPassword().get();
+        if (!passwordEncoder.matches(oldPassword, user.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Неверный текущий пароль");
+        }
         userRepository.delete(user);
     }
 
     public User getUserById(Long userId){
         return userRepository.findById(userId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, String.format("Пользователь с id %d не найден", userId)));
+    }
+
+    public User getUserByMail(String email){
+        return userRepository.getUserByEmail(email).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, String.format("Пользователь с email %s не найден", email)));
     }
 
 }
